+++ packages/asset-buyer/src/order_providers/standard_relayer_api_order_provider.ts
@@ -1,6 +1,7 @@
 import { HttpClient } from '@0x/connect';
 import { orderCalculationUtils } from '@0x/order-utils';
 import { APIOrder, AssetPairsResponse, OrderbookResponse } from '@0x/types';
+import { BigNumber } from '@0x/utils';
 import * as _ from 'lodash';
 
 import {
@@ -25,12 +26,14 @@ export class StandardRelayerAPIOrderProvider implements OrderProvider {
     ): SignedOrderWithRemainingFillableMakerAssetAmount[] {
         const result = _.map(apiOrders, apiOrder => {
             const { order, metaData } = apiOrder;
-            // calculate remainingFillableMakerAssetAmount from api metadata, else assume order is completely fillable
-            const remainingFillableTakerAssetAmount = _.get(
-                metaData,
-                'remainingTakerAssetAmount',
-                order.takerAssetAmount,
-            );
+            // The contents of metaData is not explicity defined in the spec
+            // We check for remainingTakerAssetAmount as a string and use this value if populated
+            const metaDataRemainingTakerAssetAmount = _.get(metaData, 'remainingTakerAssetAmount') as
+                | string
+                | undefined;
+            const remainingFillableTakerAssetAmount = metaDataRemainingTakerAssetAmount
+                ? new BigNumber(metaDataRemainingTakerAssetAmount)
+                : order.takerAssetAmount;
             const remainingFillableMakerAssetAmount = orderCalculationUtils.getMakerFillAmount(
                 order,
                 remainingFillableTakerAssetAmount,