+++ contracts/exchange-forwarder/CHANGELOG.md
@@ -5,6 +5,22 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v3.0.3 - _May 24, 2019_
+
+    * Dependencies updated
+
+## v3.0.2 - _May 15, 2019_
+
+    * Dependencies updated
+
+## v3.0.1 - _May 14, 2019_
+
+    * Dependencies updated
+
+## v3.0.0 - _May 10, 2019_
+
+    * Update contracts to solidity ^0.5.5 and unpin dependencies (#1796)
+
 ## v2.1.1 - _April 11, 2019_
 
     * Dependencies updated