+++ packages/sol-tracing-utils/CHANGELOG.md
@@ -5,6 +5,14 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v6.0.12 - _May 24, 2019_
+
+    * Dependencies updated
+
+## v6.0.11 - _May 10, 2019_
+
+    * Dependencies updated
+
 ## v6.0.10 - _April 11, 2019_
 
     * Dependencies updated